feat: keep the resolve internal hosts as a separate utils function
